fixing crossorigin errors

+++ services/insertMarks/src/main/java/com/stage/insertMarks/repository/GradeRepository.java
@@ -3,6 +3,6 @@ package com.stage.insertMarks.repository;
 import com.stage.insertMarks.entity.Grade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GradeRepository extends JpaRepository<Grade, Long> {
+public interface GradeRepository extends JpaRepository<Grade, Integer> {
 }
