+++ services/insertMarks/src/main/java/com/stage/insertMarks/repository/PlanEtudeRepository.java
@@ -3,6 +3,6 @@ package com.stage.insertMarks.repository;
 import com.stage.insertMarks.entity.PlanEtude;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanEtudeRepository extends JpaRepository<PlanEtude, Long> {
+public interface PlanEtudeRepository extends JpaRepository<PlanEtude, Integer> {
 }
